Revert "Rename resource file to not collide with std lib"
