+++ karabo/resource.py
